backend/pyproject.toml
[project]
name = "anantra-lms-backend"
version = "0.1.0"
description = "Backend for Anantra LMS Speech-Driven Lessons"
requires-python = ">=3.10"
dependencies = [
    "robyn>=0.68.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.0",
    "supabase>=2.0.0",
    "firebase-admin>=6.0.0",
    "redis>=5.0.0",
    "edge-tts>=6.0.0",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "google-genai>=1.0.0",
    "pypdf>=4.0.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
# vapi-python doesn't support Windows (requires daily-python which is Linux/macOS only)
vapi = ["vapi-python>=0.1.0"]

[dependency-groups]
dev = [
    "pre-commit>=3.5.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "flake8-docstrings>=1.7.0",
    "flake8-bugbear>=24.0.0",
    "isort>=5.13.0",
    "detect-secrets>=1.4.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "pyright>=1.1.390",
]
